+++ javascript/data-structures/stacksAndQueues/__tests__/stacks-queues.test.js
@@ -1,6 +1,7 @@
 'use strict';
 
-const { Stack , Queue  } = require('../lib/stacks-and-queues.js');
+const Stack = require('../lib/stack.js')
+const Queue = require('../lib/queue.js')
 
 describe("Stacks and Queues Testing", () => {
     it('can make an empty stack', () => {
@@ -16,10 +17,10 @@ describe("Stacks and Queues Testing", () => {
         let third = 3;
 
         list.push(first);
-        expect(list.head.value).toEqual(1);
+        expect(list.top.value).toEqual(1);
 
         list.push(second);
-        expect(list.head.next.value). toEqual(2);
+        expect(list.top.next.value). toEqual(2);
 
         list.push(third);
         expect(list.top.next.next.value).toEqual(3);
@@ -32,13 +33,13 @@ describe("Stacks and Queues Testing", () => {
         let second = 2;
 
         list.push(first);
-        expect(list.head.value).toEqual(1);
+        expect(list.top.value).toEqual(1);
 
         list.push(second);
-        expect(list.head.value). toEqual(2);
+        expect(list.top.value). toEqual(2);
 
         list.pop();
-        expect(list.head.value). toEqual(1);
+        expect(list.top.value). toEqual(1);
         
     })
 
@@ -73,7 +74,7 @@ describe("Stacks and Queues Testing", () => {
         list.push(fourth);
         let peeker = list.peek()
         expect(peeker).toEqual(2);
-        expect(listenerCount.top).toEqual(1)
+        expect(list.top).toEqual(1)
         
     })
 
@@ -92,13 +93,13 @@ describe("Stacks and Queues Testing", () => {
         let second = 2;
         let third = 3;
 
-        list.push(first);
+        list.enqueue(first);
         expect(list.front.value).toEqual(1);
 
-        list.push(second);
+        list.enqueue(second);
         expect(list.front.next.value). toEqual(2);
 
-        list.push(third);
+        list.enqueue(third);
         expect(list.front.next.next.value).toEqual(3);
         
     })
@@ -108,13 +109,13 @@ describe("Stacks and Queues Testing", () => {
         let first = 1;
         let second = 2;
 
-        list.push(first);
+        list.enqueue(first);
         expect(list.front.value).toEqual(1);
 
-        list.push(second);
+        list.enqueue(second);
         expect(list.front.value). toEqual(2);
 
-        list.pop();
+        list.dequeue();
         expect(list.front.value). toEqual(1);
         
     })
@@ -126,16 +127,16 @@ describe("Stacks and Queues Testing", () => {
         let third = 3;
         let fourth = 4;
 
-        list.push(first);
-        list.push(second);
-        list.push(fourth);
-        list.push(third);
+        list.enqueue(first);
+        list.enqueue(second);
+        list.enqueue(fourth);
+        list.enqueue(third);
 
-        list.pop();
-        list.pop();
-        list.pop();
+        list.dequeue();
+        list.dequeue();
+        list.dequeue();
         expect(list.front.value). toEqual(1);
-        list.pop();
+        list.dequeue();
         expect(list.front). toEqual(null);
     })
 
@@ -145,157 +146,12 @@ describe("Stacks and Queues Testing", () => {
         let second = 2;
         let fourth = 4
 
-        list.push(first);
-        list.push(second);
-        list.push(fourth);
+        list.enqueue(first);
+        list.enqueue(second);
+        list.enqueue(fourth);
         let peeker = list.peek()
         expect(peeker).toEqual(2);
         expect(listenerCount.front).toEqual(1)
         
     })
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-    it('The head property will properly point to the first node in the linked list', () => {
-        let list = new LL();
-        let first = 1;
-        let second = 2;
-
-        list.append(first);
-        expect(list.head.value).toEqual(1);
-
-        list.append(second);
-        expect(list.head.value).toEqual(1);
-        
-    })
-
-    it('Will return true when searching for a value in the linked list that does exist', () => {
-        let list = new LL();
-        let first = 1;
-        let second = 2;
-        let third = 3;
-
-        list.append(first);
-        list.append(second);
-        list.append(third);
-        let result = list.includes(3);
-        expect(result).toEqual(false);
-
-    })
-
-    it('Will return false when searching for a value in the linked list that does not exist', () => {
-        let list = new LL();
-        let first = 1;
-        let second = 2;
-
-        list.append(first);
-        list.append(second);
-        let result = list.includes(3);
-        expect(result).toEqual(false);
-
-    })
-
-    it('Can properly return a collection of all the values that exist in the linked list', () => {
-        let list = new LL();
-        let first = 1;
-        let second = 2;
-        let fourth = 44;
-        list.append(first);
-        list.append(second);
-        list.append(fourth);
-        let testing = list.toString();
-        expect(testing).toEqual('{ 1 } -> { 2 } -> { 44 } -> NULL');
-    })
-
-    it('Can properly delete a node of a given value', () => {
-        let list = new LL();
-        let first = 1;
-        let second = 2;
-        let fourth = 44;
-        list.append(first);
-        list.append(second);
-        list.append(fourth);
-
-        list.delete(2);
-        expect(list.head.next.value).toEqual(44);
-    })
-
-    it('Can locate the value of nodes nth from the tail of the LL', () => {
-        let nthList = new LL();
-        let first = 1;
-        let second = 2;
-        let fourth = 44;
-        nthList.append(first);
-        nthList.append(second);
-        nthList.append(fourth);
-        let testing = nthList.nthFromEnd(0);
-        expect(testing).toEqual(44);
-        let testing2 = nthList.nthFromEnd(2);
-        expect(testing2).toEqual(1);
-    })
-
-    it('Can run the edge cases', () => {
-        let edgeList = new LL();
-        let first = 1;
-        let second = 2;
-        let fourth = 44;
-        edgeList.append(first);
-        edgeList.append(second);
-        edgeList.append(fourth);
-        let testing = edgeList.nthFromEnd(6);
-        expect(testing).toEqual(null);
-
-        let testing2 = edgeList.nthFromEnd(2);
-        expect(testing2).toEqual(1);
-
-        let testing3 = edgeList.nthFromEnd(-1);
-        expect(testing3).toEqual(null);
-
-        let newLL = new LL();
-        newLL.append(2);
-        let one = newLL.nthFromEnd(1);
-        expect(one).toEqual(null);
-
-        newLL.append(3);
-        newLL.append(4);
-        newLL.append(67);
-        newLL.append(7);
-        newLL.append(35);
-
-        let middle = newLL.nthFromEnd(3);
-        expect(middle).toEqual(4)
-
-
-        
-
-    })
-
 })
