+++ javascript/data-structures/stacksAndQueues/lib/node.js
@@ -1,0 +1,11 @@
+'use strict';
+
+
+class Node {
+    constructor (value){
+        this.value = value, 
+        this.next = null;
+    }
+}
+
+module.exports = Node
